% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pptx_content_dimensions.R
\name{pptx_content_dimensions}
\alias{pptx_content_dimensions}
\title{Extract Content Dimension from pptx document}
\usage{
pptx_content_dimensions(file)
}
\arguments{
\item{file}{Filepath to pptx document}
}
\value{
Named list
}
\description{
Extract Content Dimension from pptx document
}
